+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add `EndpointConfig::introspectionConfig()` to pass options to `Introspection::getIntrospectionQuery()`, defaulting to `includeDeprecated: true`
+
 ## v1.1.2
 
 ### Fixed